Refactor parser and add service layer
