+++ messages.py
@@ -1,0 +1,47 @@
+MESSAGES = {
+    'ADD_DESCRIPTION': (
+        "➕ **Добавление участника**\n\n"
+        "🔴 **ОБЯЗАТЕЛЬНЫЕ ПОЛЯ:**\n"
+        "- Имя (рус) - полное имя на русском языке\n"
+        "- Пол - M/F, муж/жен, мужской/женский\n"
+        "- Размер - XS, S, M, L, XL, XXL\n"
+        "- Церковь - название церкви\n"
+        "- Роль - CANDIDATE/кандидат или TEAM/команда\n"
+        "- Департамент - обязательно для TEAM (Worship, Media, Kitchen и т.д.)\n\n"
+        "🟡 **ОПЦИОНАЛЬНЫЕ ПОЛЯ:**\n"
+        "- Имя (англ) - полное имя на английском\n"
+        "- Город - город проживания\n"
+        "- Кто подал - имя подавшего заявку\n"
+        "- Контакты - телефон или email\n\n"
+        "📝 Ниже готовый шаблон для заполнения ⬇️"
+    ),
+    'ADD_TEMPLATE': (
+        "Имя (рус): \n"
+        "Имя (англ): \n"
+        "Пол: \n"
+        "Размер: \n"
+        "Церковь: \n"
+        "Роль: \n"
+        "Департамент: \n"
+        "Город: \n"
+        "Кто подал: \n"
+        "Контакты:"
+    ),
+    'CONFIRMATION_INTRO': (
+        "🔍 Вот что удалось распознать\n"
+        "Ниже готовый темплейт с вашими данными ⬇️\n\n"
+        "✅ **Всё правильно?**\n"
+        "- Отправьте **ДА** для сохранения\n"
+        "- Отправьте **НЕТ** для отмены\n"
+        "- Или пришлите исправленный темплейт\n\n"
+        "❌ /cancel для полной отмены"
+    ),
+    'VALIDATION_ERRORS': {
+        'FullNameRU': 'Не указано имя',
+        'Gender': 'Пол должен быть M или F',
+        'Church': 'Не указана церковь',
+        'Role': 'Не указана роль',
+        'Department': 'Для роли TEAM необходимо указать департамент',
+        'Size': 'Недопустимый размер одежды',
+    },
+}
