+++ utils/validators.py
@@ -1,4 +1,5 @@
 from typing import Dict
+from messages import MESSAGES
 
 VALID_SIZES = [
     'XS', 'EXTRA SMALL', 'EXTRASMALL',
@@ -15,8 +16,22 @@ def validate_size(size: str) -> bool:
 
 
 def validate_participant_data(data: Dict) -> (bool, str):
-    if data.get('Size') and not validate_size(data['Size']):
-        return False, 'Недопустимый размер одежды'
+    if not data.get('FullNameRU'):
+        return False, MESSAGES['VALIDATION_ERRORS']['FullNameRU']
+
+    if data.get('Gender') not in ('M', 'F'):
+        return False, MESSAGES['VALIDATION_ERRORS']['Gender']
+
+    if not data.get('Church'):
+        return False, MESSAGES['VALIDATION_ERRORS']['Church']
+
+    if not data.get('Role'):
+        return False, MESSAGES['VALIDATION_ERRORS']['Role']
+
     if data.get('Role') == 'TEAM' and not data.get('Department'):
-        return False, 'Для роли TEAM необходимо указать департамент'
+        return False, MESSAGES['VALIDATION_ERRORS']['Department']
+
+    if data.get('Size') and not validate_size(data['Size']):
+        return False, MESSAGES['VALIDATION_ERRORS']['Size']
+
     return True, ''
